Update Ex_2.2(bmi calculator ).py

+++ Day_02/Ex_2.2(bmi calculator ).py
@@ -4,15 +4,13 @@ height = input("write your height(m)  :  ")
 
 #input function always give string so convert in integer 
 
-X=int(weight)
-Y=int(height)
+X=float(weight)
+Y=float(height)
 
-bmi=str(X/Y**2)
+bmi= str(X/Y**2)
 
 # we need to convert x/y**2 in str because we can't cacenate integer with string see next step'
 
 print("Your body mass index(bmi) is = " +bmi)
 
-#alternate
-#bmi=X/Y**2
-#print("Your body amss index(bmi) is = " + str(bmi))
+
